browse and filter tests fully automated
